index mobile page styling complete
Styling for the mobile index page is complete. The header and footer took the most time here, and now that they're done, the other pages should go much quicker.

Important to note that I'm on the wrong branch, but committing anyway because the old index branch didn't have the admin php, so let's avoid that branch altogether.