progress
Been working on this for 4 hours, going to step back for a bit. Left off at the events section and the contact form which is currently hidden. Remove hidden class if you want to see the progress. Got info from the actual LRG website, tried to merge with current design.